copy_file: drop a redundant u64 conversion

+++ src/runtime/webcore/blob/copy_file.rs
@@ -506,7 +506,7 @@ impl CopyFile {
         // source's st_size. A source fd whose position is past 0 supplies fewer
         // bytes than that, so trim the destination to what was copied.
         if !unknown_size
-            && total_written < u64::from(self.max_length)
+            && total_written < self.max_length
             && matches!(
                 self.destination_file_store.pathlike,
                 PathOrFileDescriptor::Path(_)
